change concept - users wont be able to manually save the history because i feel it doesnt make sense to have partial history anyway - history gets saved temporarily when viewing it, so we can just pass the history entity id as an argument rather than the entire domain model - timeline saver renamed to history detail, which can be accessed through the main activity

+++ app/src/main/java/com/dgnt/quickScoreboardCreator/ui/main/historylist/HistoryListViewModel.kt
@@ -34,13 +34,13 @@ class HistoryListViewModel @Inject constructor(
 
     private var deletedHistoryList: MutableList<HistoryEntity> = mutableListOf()
 
-    fun onEdit(id: Long) = sendUiEvent(UiEvent.HistoryDetails(id))
-    fun onLaunch(id: Long) {
+    fun onEdit(id: Int) = sendUiEvent(UiEvent.HistoryDetails(id))
+    fun onLaunch(id: Int) {
 
         //TODO launch timeline viewer
     }
 
-    fun onDelete(id: Long) = viewModelScope.launch {
+    fun onDelete(id: Int) = viewModelScope.launch {
         historyEntityList.first().find { entity ->
             entity.id == id
         }?.let {
